src/main/java/com/amazon/service/CategoryService.java
package com.amazon.service;

import com.amazon.domain.Category;
import com.amazon.repository.CategoryRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor(onConstructor = @__(@Autowired))
public class CategoryService {

    private final CategoryRepository repository;

    public List<Category> getAllCategories() {
        return repository.findAll();
    }

    public List<Category> getAllParentCategories(){ return repository.findByParentCategoryNull();}

}
